Correct SetActionUseMovePlayer's parameter types; retire the last stale extern
The previous commit left src/overlay_29_022F0EDC.c carrying a provisional
`extern s32 SetActionUseMovePlayer();` because replacing it with the real header
broke three modules. The cause was not the return type, which is void, but the
parameters: the function was landed as (struct action_data*, u8, u8), and a u8
parameter makes every caller truncate.

SetLeaderAction's call site in retail passes GetTeamMemberIndex's result
unchanged and sign-extends the move index to 16 bits:

    mov r1, r0
    mov r2, r4, lsl #0x10
    mov r2, r2, asr #0x10
    bl SetActionUseMovePlayer

so the parameters are s32 and s16. The callee's own body cannot distinguish
u8, s16 or s32 -- it stores both with strb, and all three score 0 against it --
which is why the wrong types went unnoticed when it was landed.

With (struct action_data*, s32, s16) both functions match and the provisional
extern is deleted in favour of the header, leaving one declaration in the tree.

Authored by Claude (Opus 5) under human direction. Confirmed by a matching
build: build/pmdsky.us/pmdsky.us.nds: OK.

+++ include/overlay_29_022EBC50.h
@@ -9,6 +9,6 @@
 void SetActionRegularAttack(struct action_data *monster_action, u8 direction);
 // Sets a monster's action to action::ACTION_STRUGGLE, with a specified direction.
 void SetActionStruggle(struct action_data *monster_action, u8 direction);
-void SetActionUseMovePlayer(struct action_data* action, u8 monster_index, u8 move_index);
+void SetActionUseMovePlayer(struct action_data* action, s32 monster_index, s16 move_index);
 
 #endif //PMDSKY_OVERLAY_29_022EBC50_H
